refactor(web): narrow the empty-state work to the notifications bell
The Activity page empty state goes back to what main had. That page is a
candidate for removal on its own, so redesigning it was scope nobody asked
for, and the bell popover is the surface the work was for.

What survives is the bell: an icon well, the headline, and one recipe that
opens a fresh conversation seeded with the prompt for a morning briefing,
dismissing the panel first. No description line and no secondary action,
because the panel already names itself "Notifications" above it. A failed
load keeps its own message.

`empty-state-scene` keeps only the pieces the bell renders. The preview
layer, the recipe grid's column choices, and the scene's description,
actions, and page density all lost their last consumer with the page
variant.

Restores `homeFeedList.emptyTitle` and `homeFeedList.emptyBody`, which the
reverted page reads again.
