+++ clients/web/src/domains/home/home-feed-list.tsx
@@ -1,6 +1,7 @@
-import { ChevronRight } from "lucide-react";
+import { Bell, ChevronRight } from "lucide-react";
 import { useState } from "react";
 
+import { PageEmptyState } from "@/components/page-empty-state";
 import { useTranslation } from "@/i18n";
 
 import type {
@@ -9,7 +10,6 @@ import type {
   FeedItemStatus,
 } from "@vellumai/assistant-api";
 import { Collapsible, Typography } from "@vellumai/design-library";
-import { NotificationsEmptyState } from "./components/notifications-empty-state";
 import { HomeFeedFilterBar } from "./home-feed-filter-bar";
 import { HomeRecapRow } from "./home-recap-row";
 import type { FeedTimeGroup } from "./utils";
@@ -135,7 +135,11 @@ export function HomeFeedList({
             {t("homeFeedList.noMatches")}
           </Typography>
         ) : (
-          <NotificationsEmptyState />
+          <PageEmptyState
+            icon={Bell}
+            title={t("homeFeedList.emptyTitle")}
+            description={t("homeFeedList.emptyBody")}
+          />
         )
       ) : (
         [...grouped.entries()].map(([group, groupItems]) => (
