+++ clients/web/src/i18n/locales/en/home.json
@@ -30,6 +30,8 @@
     "yesterday": "Yesterday",
     "older": "Older",
     "noMatches": "No items match the selected filter.",
+    "emptyTitle": "No notifications yet",
+    "emptyBody": "Updates and activity from your assistant will appear here.",
     "earlier": "Earlier ({count, number})",
     "dismissed": "Dismissed ({count, number})"
   },
@@ -84,32 +86,11 @@
       "leftUnverified": "Left unverified"
     }
   },
-  "notificationsEmptyState": {
-    "title": "Nothing to report yet.",
-    "description": "I'll post here when a schedule runs, a reminder fires, or something needs your attention.",
-    "previewLabel": "Example",
-    "briefingExampleTitle": "Morning briefing",
-    "briefingExampleSummary": "Three meetings today, two emails waiting on a reply, and rain from 16:00.",
-    "triageExampleTitle": "Inbox triage",
-    "triageExampleSummary": "Four emails need a reply. The drafts are ready for you to review.",
-    "reminderExampleTitle": "Reminder: send the sprint recap",
-    "reminderExampleSummary": "You asked me to remind you before the team sync at 16:00.",
-    "briefingRecipeTitle": "Morning briefing",
-    "briefingRecipeMeta": "Daily",
-    "briefingRecipeDescription": "Calendar, inbox, and news, every weekday at 8:00.",
-    "triageRecipeTitle": "Inbox triage",
-    "triageRecipeMeta": "Every 2h",
-    "triageRecipeDescription": "Every 2 hours, flag what needs a reply and draft it.",
-    "seeSchedules": "See all schedules"
-  },
   "notificationsBellEmptyState": {
     "title": "Nothing yet.",
     "briefingRecipeTitle": "Set up a morning briefing",
     "briefingRecipeMeta": "Daily",
-    "briefingRecipeDescription": "Your first notification, tomorrow at 8:00."
-  },
-  "notificationRecipeCard": {
-    "briefingPrompt": "Set up a morning briefing for me: every weekday at 8:00, check my calendar, my inbox, and the news, then post a short summary to my notifications.",
-    "triagePrompt": "Set up a schedule that runs every 2 hours: go through my inbox, flag the emails that need a reply, and draft each reply for me to review."
+    "briefingRecipeDescription": "Your first notification, tomorrow at 8:00.",
+    "briefingRecipePrompt": "Set up a morning briefing for me: every weekday at 8:00, check my calendar, my inbox, and the news, then post a short summary to my notifications."
   }
 }
