+++ clients/web/src/i18n/locales/ru/home.json
@@ -30,6 +30,8 @@
     "yesterday": "Вчера",
     "older": "Ранее",
     "noMatches": "Нет элементов по выбранному фильтру.",
+    "emptyTitle": "Уведомлений пока нет",
+    "emptyBody": "Здесь появятся обновления и активность ассистента.",
     "earlier": "Ранее ({count, number})",
     "dismissed": "Скрытые ({count, number})"
   },
@@ -84,32 +86,11 @@
       "leftUnverified": "Оставлен без проверки"
     }
   },
-  "notificationsEmptyState": {
-    "title": "Пока не о чем сообщить.",
-    "description": "Я напишу здесь, когда сработает расписание или напоминание либо когда что-то потребует вашего внимания.",
-    "previewLabel": "Пример",
-    "briefingExampleTitle": "Утренняя сводка",
-    "briefingExampleSummary": "Сегодня три встречи, два письма ждут ответа и дождь после 16:00.",
-    "triageExampleTitle": "Разбор входящих",
-    "triageExampleSummary": "Четыре письма ждут ответа. Черновики готовы к проверке.",
-    "reminderExampleTitle": "Напоминание: отправить итоги спринта",
-    "reminderExampleSummary": "Вы просили напомнить перед командной встречей в 16:00.",
-    "briefingRecipeTitle": "Утренняя сводка",
-    "briefingRecipeMeta": "Ежедневно",
-    "briefingRecipeDescription": "Календарь, почта и новости каждый будний день в 8:00.",
-    "triageRecipeTitle": "Разбор входящих",
-    "triageRecipeMeta": "Каждые 2 ч",
-    "triageRecipeDescription": "Каждые 2 часа отмечаю письма, которые ждут ответа, и готовлю черновик.",
-    "seeSchedules": "Все расписания"
-  },
   "notificationsBellEmptyState": {
     "title": "Пока ничего.",
     "briefingRecipeTitle": "Настроить утреннюю сводку",
     "briefingRecipeMeta": "Ежедневно",
-    "briefingRecipeDescription": "Первое уведомление придёт завтра в 8:00."
-  },
-  "notificationRecipeCard": {
-    "briefingPrompt": "Настрой мне утреннюю сводку: каждый будний день в 8:00 проверяй мой календарь, входящие и новости, а затем публикуй краткую сводку в моих уведомлениях.",
-    "triagePrompt": "Настрой расписание, которое запускается каждые 2 часа: просматривай мои входящие, отмечай письма, которые ждут ответа, и готовь черновик ответа на каждое, чтобы я мог их проверить."
+    "briefingRecipeDescription": "Первое уведомление придёт завтра в 8:00.",
+    "briefingRecipePrompt": "Настрой мне утреннюю сводку: каждый будний день в 8:00 проверяй мой календарь, входящие и новости, а затем публикуй краткую сводку в моих уведомлениях."
   }
 }
